+++ include/uq.hpp
@@ -10,7 +10,7 @@ class UncertaintyQuantifier {
 public:
     UncertaintyQuantifier(int reservoir_size = 100) : reservoir_size(reservoir_size) {}
 
-    void update_reservoir(double score) {
+    void update_reservoir(float score) {
         if (scores.size() < reservoir_size) {
             scores.push_back(score);
         } else {
@@ -18,33 +18,33 @@ public:
         }
     }
 
-    double get_conformal_threshold(double alpha = 0.1) {
-        if (scores.empty()) return 1.0;
-        std::vector<double> sorted_scores = scores;
+    float get_conformal_threshold(float alpha = 0.1f) {
+        if (scores.empty()) return 1.0f;
+        std::vector<float> sorted_scores = scores;
         std::sort(sorted_scores.begin(), sorted_scores.end());
-        int idx = static_cast<int>((1.0 - alpha) * sorted_scores.size());
+        int idx = static_cast<int>((1.0f - alpha) * sorted_scores.size());
         return sorted_scores[std::min(idx, (int)sorted_scores.size() - 1)];
     }
 
-    double compute_total_uncertainty(const std::vector<Eigen::VectorXd>& ensemble_outputs) {
-        if (ensemble_outputs.empty()) return 0.0;
+    float compute_total_uncertainty(const std::vector<Eigen::VectorXf>& ensemble_outputs) {
+        if (ensemble_outputs.empty()) return 0.0f;
         
-        Eigen::VectorXd mean = Eigen::VectorXd::Zero(ensemble_outputs[0].size());
+        Eigen::VectorXf mean = Eigen::VectorXf::Zero(ensemble_outputs[0].size());
         for (const auto& out : ensemble_outputs) mean += out;
-        mean /= ensemble_outputs.size();
+        mean /= (float)ensemble_outputs.size();
 
-        double variance = 0.0;
+        float variance = 0.0f;
         for (const auto& out : ensemble_outputs) {
             variance += (out - mean).squaredNorm();
         }
-        return variance / ensemble_outputs.size();
+        return variance / (float)ensemble_outputs.size();
     }
 
     void save(std::ostream& os) const {
         size_t size = scores.size();
         os.write((char*)&size, sizeof(size));
         if (size > 0) {
-            os.write((char*)scores.data(), size * sizeof(double));
+            os.write((char*)scores.data(), size * sizeof(float));
         }
     }
 
@@ -53,13 +53,14 @@ public:
         is.read((char*)&size, sizeof(size));
         scores.resize(size);
         if (size > 0) {
-            is.read((char*)scores.data(), size * sizeof(double));
+            is.read((char*)scores.data(), size * sizeof(float));
         }
     }
 
 private:
     int reservoir_size;
-    std::vector<double> scores;
+    std::vector<float> scores;
 };
 
 #endif // UQ_HPP
+
